Opprettet idenservice og utils klasse
