+++ apps/testnorge-bisys/src/main/java/no/nav/registre/bisys/provider/request/SyntetiserBisysRequest.java
@@ -1,8 +1,6 @@
-package no.nav.registre.bisys.provider.requests;
+package no.nav.registre.bisys.provider.request;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Value;
